new-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pymupdf==1.23.25
sentence-transformers==2.5.1
ollama==0.1.6
python-dotenv==1.0.1
sqlalchemy==2.0.28
aiofiles==23.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
